Change pass score type to float

+++ server/src/quizzes/dto/create-quiz.input.ts
@@ -1,4 +1,4 @@
-import { InputType, Field, Int } from '@nestjs/graphql';
+import { InputType, Field, Float } from '@nestjs/graphql';
 import { IsInt, Length, Max, Min } from 'class-validator';
 import { CreateQuestionBaseInput } from 'src/questions/dto';
 
@@ -17,7 +17,7 @@ export class CreateQuizInput {
   @Field()
   level: string;
 
-  @Field(() => Int, { defaultValue: 70 })
+  @Field(() => Float, { defaultValue: 70 })
   @IsInt()
   @Min(50)
   @Max(100)
